Move feeding to its own component and fixed to work with institutes
